python/pyproject.toml
[project]
name = "wt"
version = "0.1.0"
description = "Git worktree orchestrator for parallel agent sessions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
]

[project.scripts]
wt = "wt.cli:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wt"]
